Changed <ins> Syntax
[brackets] replaced with +++pluses+++
Provides a logical counterpart to ---false--- (<del>) and allows square brackets to be used for other purposes.
According to the HTML spec this [<ins>] was intended primarily for use in marking up versioning of a document.
Brackets are used, among other things, in mathematics and to enclose notes by someone other than the original other, especially in quoted text.
More information here:
<http://en.wikipedia.org/wiki/Bracket#Square_brackets_.5B_.5D>